requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.19
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
openai>=1.35.0
gTTS>=2.5.1
python-multipart>=0.0.9
